Simplify package requirements
Currently, both the README and the Build-Depends of the Debian package
require versions of libraries as available in Debian Buster; however,
vonsh does not really require them, i.e. it builds fine in Debian
Stretch too. Also, some of the packages required for the Debian build
are part of the base build-essential set, and thus no need to be
specified (unless a specific version is required, which is not the case).

Hence, simplify the list of requirements:
- do not mention libc6 as requirement: it is the C library, always
  available
- mention the 3 SDL modules by name, not by Debian package name
- remove libc6-dev, gcc, g++, make, and dpkg-dev from the build
  dependencies: any version of them works just fine, and they are part
  of build-essential
- mention build-essential only in the README, so users know what to
  install
- lower the version of debhelper to 9, as it is the compat level
  requested by debian/compat
- drop the version of libsdl2-dev, libsdl2-image-dev, and
  libsdl2-mixer-dev, as older versions work just fine